# packages/api-contracts/openapi/wishpool.yaml
openapi: 3.1.0
info:
  title: WishPool API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    WishPool core API contract for mobile apps, parent web, admin web, and backend services.
    All write endpoints are command-oriented and should be called with an Idempotency-Key header.
servers:
  - url: http://localhost:8080
    description: Local core-api
security:
  - bearerAuth: []
...
paths:
...
  /task-templates:
    get:
      tags: [Task Templates]
      operationId: listTaskTemplates
      summary: List task templates for the current family.
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task templates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskTemplate'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Task Templates]
      operationId: createTaskTemplate
      summary: Create a task template.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskTemplateRequest'
      responses:
        '201':
          description: Template created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskTemplate'
        default:
          $ref: '#/components/responses/Error'
  /plans:
    post:
      tags: [Plans]
      operationId: saveWeeklyPlan
      summary: Create or update a weekly plan.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveWeeklyPlanRequest'
      responses:
        '200':
          description: Weekly plan saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        default:
          $ref: '#/components/responses/Error'
  /plans/{planId}:
    get:
      tags: [Plans]
      operationId: getWeeklyPlan
      summary: Get a weekly plan.
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Weekly plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlan'
        default:
          $ref: '#/components/responses/Error'
  /children/{childId}/today:
    get:
      tags: [Tasks]
      operationId: getToday
      summary: Get child today dashboard data.
      parameters:
        - $ref: '#/components/parameters/ChildId'
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Today snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodaySnapshot'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{taskId}/skip:
    post:
      tags: [Tasks]
      operationId: skipTask
      summary: Mark a task as skipped.
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkipTaskRequest'
      responses:
        '200':
          description: Task skipped.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskInstance'
        default:
          $ref: '#/components/responses/Error'
  /tasks/{taskId}/postpone:
    post:
      tags: [Tasks]
      operationId: postponeTask
      summary: Postpone a task to another date.
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostponeTaskRequest'
      responses:
        '200':
          description: Task postponed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskInstance'
        default:
          $ref: '#/components/responses/Error'
...
  /submissions:
    post:
      tags: [Submissions]
      operationId: createSubmission
      summary: Create a child submission.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
      responses:
        '201':
          description: Submission created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        default:
          $ref: '#/components/responses/Error'
  /submissions/{submissionId}:
    get:
      tags: [Submissions]
      operationId: getSubmission
      summary: Get submission details.
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Submission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
        default:
          $ref: '#/components/responses/Error'
...
  /wishes:
    post:
      tags: [Wishes]
      operationId: createWish
      summary: Create a wish.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWishRequest'
      responses:
        '201':
          description: Wish created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        default:
          $ref: '#/components/responses/Error'
  /wishes/{wishId}/activate:
    post:
      tags: [Wishes]
      operationId: activateWish
      summary: Activate a wish.
      parameters:
        - $ref: '#/components/parameters/WishId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Wish activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        default:
          $ref: '#/components/responses/Error'
  /wishes/{wishId}:
    get:
      tags: [Wishes]
      operationId: getWish
      summary: Get a wish.
      parameters:
        - $ref: '#/components/parameters/WishId'
      responses:
        '200':
          description: Wish.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        default:
          $ref: '#/components/responses/Error'
  /wishes/{wishId}/redeem:
    post:
      tags: [Wishes]
      operationId: redeemWish
      summary: Redeem an unlocked wish.
      parameters:
        - $ref: '#/components/parameters/WishId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemWishRequest'
      responses:
        '200':
          description: Wish redeemed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishRedemption'
        default:
          $ref: '#/components/responses/Error'
  /children/{childId}/wishes/current:
    get:
      tags: [Wishes]
      operationId: getCurrentWish
      summary: Get the current wish for a child.
      parameters:
        - $ref: '#/components/parameters/ChildId'
      responses:
        '200':
          description: Current wish.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wish'
        default:
          $ref: '#/components/responses/Error'
  /children/{childId}/wishes:
    get:
      tags: [Wishes]
      operationId: listChildWishes
      summary: List wishes for a child.
      parameters:
        - $ref: '#/components/parameters/ChildId'
        - name: weekId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Child wishes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wish'
        default:
          $ref: '#/components/responses/Error'
...
  /room/state:
    get:
      tags: [Room]
      operationId: getRoomState
      summary: Get room state for a child.
      parameters:
        - name: childId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Room state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomState'
        default:
          $ref: '#/components/responses/Error'
  /room/items/{itemId}/arrange:
    post:
      tags: [Room]
      operationId: arrangeRoomItem
      summary: Arrange a room item.
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArrangeRoomItemRequest'
      responses:
        '200':
          description: Room item arranged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomItem'
        default:
          $ref: '#/components/responses/Error'
  /sync/pull:
    get:
      tags: [Sync]
      operationId: pullSyncEvents
      summary: Pull family events after a sequence number.
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: afterSeq
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 500
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Family events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPullResponse'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WishId:
      name: wishId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
...
    ChildProfile:
      type: object
      required: [id, familyId, nickname, roomTheme, status]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        nickname:
          type: string
        birthYear:
          anyOf:
            - type: integer
            - type: 'null'
        avatarAsset:
          anyOf:
            - type: string
            - type: 'null'
        roomTheme:
          type: string
        status:
          type: string
          enum: [active, archived]
...
    TaskCategory:
      type: string
      enum: [study, reading, exercise, habit, custom]
    SubmissionType:
      type: string
      enum: [photo, audio, video, manual]
    TaskTemplate:
      type: object
      required: [id, familyId, title, category, submissionType]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        title:
          type: string
        category:
          $ref: '#/components/schemas/TaskCategory'
        submissionType:
          $ref: '#/components/schemas/SubmissionType'
        description:
          anyOf:
            - type: string
            - type: 'null'
        targetText:
          anyOf:
            - type: string
            - type: 'null'
        defaultDurationSec:
          anyOf:
            - type: integer
            - type: 'null'
    CreateTaskTemplateRequest:
      type: object
      required: [familyId, title, category, submissionType]
      properties:
        familyId:
          type: string
          format: uuid
        title:
          type: string
        category:
          $ref: '#/components/schemas/TaskCategory'
        submissionType:
          $ref: '#/components/schemas/SubmissionType'
        description:
          type: string
        targetText:
          type: string
        defaultDurationSec:
          type: integer
    SaveWeeklyPlanRequest:
      type: object
      required: [familyId, childId, weekId, startDate, endDate, rewardMode, rules]
      properties:
        familyId:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        weekId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        rewardMode:
          type: string
          enum: [flexible, strict]
        wishId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        rules:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyPlanRuleInput'
    WeeklyPlanRuleInput:
      type: object
      required: [title, category, submissionType, weekdays, isCore, requireReview]
      properties:
        taskTemplateId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        title:
          type: string
        category:
          $ref: '#/components/schemas/TaskCategory'
        submissionType:
          $ref: '#/components/schemas/SubmissionType'
        description:
          type: string
        targetText:
          type: string
        weekdays:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 7
        isCore:
          type: boolean
        requireReview:
          type: boolean
        sortOrder:
          type: integer
    WeeklyPlan:
      type: object
      required: [id, familyId, childId, weekId, startDate, endDate, rewardMode, status, rules]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        weekId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        rewardMode:
          type: string
          enum: [flexible, strict]
        status:
          type: string
          enum: [draft, active, archived]
        rules:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyPlanRule'
    WeeklyPlanRule:
      allOf:
        - $ref: '#/components/schemas/WeeklyPlanRuleInput'
        - type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    TaskInstance:
      type: object
      required: [id, familyId, childId, scheduledDate, title, category, submissionType, status, isCore, requireReview]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        scheduledDate:
          type: string
          format: date
        title:
          type: string
        category:
          $ref: '#/components/schemas/TaskCategory'
        submissionType:
          $ref: '#/components/schemas/SubmissionType'
        description:
          anyOf:
            - type: string
            - type: 'null'
        targetText:
          anyOf:
            - type: string
            - type: 'null'
        isCore:
          type: boolean
        requireReview:
          type: boolean
        status:
          type: string
          enum: [todo, submitted, ai_processing, pending_review, approved, needs_revision, skipped, expired, adjusted_by_parent]
        latestSubmissionId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
    TodaySnapshot:
      type: object
      required: [child, date, tasks]
      properties:
        child:
          $ref: '#/components/schemas/ChildProfile'
        date:
          type: string
          format: date
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskInstance'
        dailySummary:
          $ref: '#/components/schemas/DailySummary'
        currentWish:
          $ref: '#/components/schemas/Wish'
    SkipTaskRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    PostponeTaskRequest:
      type: object
      required: [newDate]
      properties:
        newDate:
          type: string
          format: date
        reason:
          type: string
...
    MediaAsset:
      type: object
      required: [id, familyId, purpose, storageKey, contentType, status]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        childId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        purpose:
          type: string
        storageKey:
          type: string
        contentType:
          type: string
        status:
          type: string
          enum: [upload_pending, uploaded, processing, ready, failed, deleted]
        downloadUrl:
          anyOf:
            - type: string
              format: uri
            - type: 'null'
    CreateSubmissionRequest:
      type: object
      required: [taskInstanceId, clientMutationId, mediaAssetIds]
      properties:
        taskInstanceId:
          type: string
          format: uuid
        clientMutationId:
          type: string
        mediaAssetIds:
          type: array
          items:
            type: string
            format: uuid
        submittedAtClient:
          type: string
          format: date-time
    Submission:
      type: object
      required: [id, familyId, childId, taskInstanceId, attemptNo, submissionType, status, submittedAt]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        taskInstanceId:
          type: string
          format: uuid
        attemptNo:
          type: integer
        submissionType:
          $ref: '#/components/schemas/SubmissionType'
        status:
          type: string
          enum: [created, media_uploaded, ai_pending, ai_processing, review_pending, approved, rejected, superseded, cancelled]
        submittedAt:
          type: string
          format: date-time
    SubmissionDetail:
      allOf:
        - $ref: '#/components/schemas/Submission'
        - type: object
          properties:
            task:
              $ref: '#/components/schemas/TaskInstance'
            media:
              type: array
              items:
                $ref: '#/components/schemas/MediaAsset'
            aiPrecheck:
              $ref: '#/components/schemas/AiPrecheck'
            review:
              $ref: '#/components/schemas/Review'
    AiPrecheck:
      type: object
      required: [id, submissionId, type, summary, flags, model]
      properties:
        id:
          type: string
          format: uuid
        submissionId:
          type: string
          format: uuid
        type:
          type: string
        summary:
          type: string
        confidence:
          anyOf:
            - type: number
              minimum: 0
              maximum: 1
            - type: 'null'
        flags:
          type: array
          items:
            type: object
            additionalProperties: true
        model:
          type: object
          additionalProperties: true
...
    Review:
      type: object
      required: [id, submissionId, taskInstanceId, decision, reviewedBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        submissionId:
          type: string
          format: uuid
        taskInstanceId:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approved, needs_revision]
        reviewedBy:
          type: string
          format: uuid
        feedback:
          anyOf:
            - $ref: '#/components/schemas/Feedback'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    Feedback:
      type: object
      properties:
        emoji:
          anyOf:
            - type: string
            - type: 'null'
        text:
          anyOf:
            - type: string
            - type: 'null'
        audioMedia:
          anyOf:
            - $ref: '#/components/schemas/MediaAsset'
            - type: 'null'
...
    CreateWishRequest:
      type: object
      required: [familyId, childId, weekId, title, requiredFragments, rewardMode]
      properties:
        familyId:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        weekId:
          type: string
        title:
          type: string
        note:
          type: string
        imageMediaId:
          type: string
          format: uuid
        requiredFragments:
          type: integer
          minimum: 1
        rewardMode:
          type: string
          enum: [flexible, strict]
    Wish:
      type: object
      required: [id, familyId, childId, weekId, title, requiredFragments, earnedFragments, rewardMode, status]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        weekId:
          type: string
        title:
          type: string
        note:
          anyOf:
            - type: string
            - type: 'null'
        imageMedia:
          anyOf:
            - $ref: '#/components/schemas/MediaAsset'
            - type: 'null'
        requiredFragments:
          type: integer
        earnedFragments:
          type: integer
        rewardMode:
          type: string
          enum: [flexible, strict]
        status:
          type: string
          enum: [draft, active, unlocked, redeemed, archived, cancelled, cancelled_by_parent]
    RedeemWishRequest:
      type: object
      required: [redeemedDate, photoMediaIds]
      properties:
        redeemedDate:
          type: string
          format: date
        photoMediaIds:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: string
            format: uuid
        parentNote:
          type: string
        childNote:
          type: string
    WishRedemption:
      type: object
      required: [id, wishId, redeemedDate, createdAt]
      properties:
        id:
          type: string
          format: uuid
        wishId:
          type: string
          format: uuid
        redeemedDate:
          type: string
          format: date
        parentNote:
          anyOf:
            - type: string
            - type: 'null'
        childNote:
          anyOf:
            - type: string
            - type: 'null'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/MediaAsset'
        createdAt:
          type: string
          format: date-time
    DailySummary:
      type: object
      properties:
        date:
          type: string
          format: date
        coreRequired:
          type: integer
        coreApproved:
          type: integer
        coreSkipped:
          type: integer
        fragmentStatus:
          type: string
...
    RoomState:
      type: object
      required: [childId, items]
      properties:
        childId:
          type: string
          format: uuid
        theme:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoomItem'
    RoomItem:
      type: object
      required: [id, childId, type, title, visible, unlockedAt]
      properties:
        id:
          type: string
          format: uuid
        childId:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        media:
          $ref: '#/components/schemas/MediaAsset'
        position:
          type: object
          additionalProperties: true
        visible:
          type: boolean
        unlockedAt:
          type: string
          format: date-time
    ArrangeRoomItemRequest:
      type: object
      required: [position]
      properties:
        position:
          type: object
          additionalProperties: true
    FamilyEvent:
      type: object
      required: [seq, familyId, type, aggregateType, aggregateId, occurredAt, payload]
      properties:
        seq:
          type: integer
          format: int64
        familyId:
          type: string
          format: uuid
        type:
          type: string
        aggregateType:
          type: string
        aggregateId:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    SyncPullResponse:
      type: object
      required: [events, latestSeq]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/FamilyEvent'
        latestSeq:
          type: integer
          format: int64
...